feat(engine): ranks, the descent, and the XP of a Couleur (T05)
A rank needs enough wins AND enough weeks under the current one. The second is
what stops a good fortnight carrying a child past what they can actually spell,
and only one rank is won per week, so two runs in an evening cannot move it
twice.

The descent is not a punishment bolted on — the simulation needed it. With no
way down every child plateaued and left out of boredom. A single win breaks the
run of failures, Blanc is the floor, and BestRank remembers the highest ever
held because the atelier is furnished from it.

Only a blind word or a boss manche earns XP: levelling comes from the harder
ways of playing, not from repeating an easy manche until it counts.

TWO BUGS, BOTH FOUND RATHER THAN GUESSED AT.

The cooldown used the week of the last level-up with zero standing for "never
levelled" — but week zero is a real week, so a Couleur levelled in it could
level again at once. It now stores the week the cooldown ENDS, whose zero value
honestly means "allowed now": the sentinel is gone rather than special-cased.

gosec flagged an int-to-int32 conversion for the third time in this package,
and the third one is the one that says what is wrong: weeks and days were int32
in the state and int in the config, so every use had to narrow one to meet the
other. The config's week and day fields now carry the type weeks and days are
counted in, and the conversions are gone from every site at once instead of
being annotated one by one.

Bead: encre-c8g

+++ engine/word_test.go
@@ -81,10 +81,10 @@ func TestAGoldWordLeftAloneForFourWeeksTarnishes(t *testing.T) {
 	cfg := engine.DefaultConfig()
 	st := &engine.WordState{Gold: true, LastPlayedW: 2}
 
-	if got := engine.Tarnish(st, 2+int32(cfg.TarnishWeeks)-1, cfg); got {
+	if got := engine.Tarnish(st, 2+cfg.TarnishWeeks-1, cfg); got {
 		t.Error("the word tarnished a week early")
 	}
-	if got := engine.Tarnish(st, 2+int32(cfg.TarnishWeeks), cfg); !got {
+	if got := engine.Tarnish(st, 2+cfg.TarnishWeeks, cfg); !got {
 		t.Errorf("the word did not tarnish after %d weeks", cfg.TarnishWeeks)
 	}
 	if !st.Tarnished {
